Create MethodOverriddenNode, add English localized string
